Move `SplitTunneling` into `ServiceEndpoint`
